Remove PP technical input tests
