+++ src/components/ProfileForm.jsx
@@ -14,6 +14,7 @@ import {
 } from 'components/inputValidations';
 import { parseUkTriggerQuery } from 'utils/parseUkTrigger';
 import { normalizeLastAction } from 'utils/normalizeLastAction';
+import { resolvePpTechnicalInputTarget } from 'utils/ppTechnicalInputTarget';
 import { patchOverlayField } from 'utils/multiAccountEdits';
 import toast from 'react-hot-toast';
 import { removeField } from './smallCard/actions';
@@ -136,7 +137,6 @@ const PHENOTYPE_FIELDS = [
   'responsibility',
 ];
 const ANTHROPOMETRY_FIELDS = ['height', 'weight', 'imt', 'clothingSize', 'shoeSize', 'breastSize'];
-const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/i;
 const appendFieldValue = (currentValue, nextValue) => {
   const normalizedNextValue = String(nextValue ?? '').trim();
   if (!normalizedNextValue) return currentValue;
@@ -153,51 +153,6 @@ const appendFieldValue = (currentValue, nextValue) => {
   return [normalizedCurrentValue, normalizedNextValue];
 };
 
-const normalizeUrlForStorage = rawValue => {
-  const trimmed = String(rawValue || '').trim();
-  if (!trimmed) return '';
-
-  if (/^https?:\/\//i.test(trimmed)) {
-    return trimmed;
-  }
-
-  if (/^www\./i.test(trimmed) || /^[\w-]+(\.[\w-]+)+/i.test(trimmed)) {
-    return `https://${trimmed}`;
-  }
-
-  return trimmed;
-};
-
-const resolvePpTechnicalInputTarget = rawValue => {
-  const trimmed = String(rawValue || '').trim();
-  if (!trimmed) return null;
-
-  const socialUrlMatchers = [
-    { fieldName: 'instagram', pattern: /(?:https?:\/\/)?(?:www\.)?instagram\.com\/([^/?#]+)/i },
-    { fieldName: 'facebook', pattern: /(?:https?:\/\/)?(?:www\.)?(?:facebook\.com|fb\.com)\/([^/?#]+)/i },
-    { fieldName: 'tiktok', pattern: /(?:https?:\/\/)?(?:www\.)?tiktok\.com\/([^/?#]+)/i },
-    { fieldName: 'linkedin', pattern: /(?:https?:\/\/)?(?:www\.)?linkedin\.com\/([^/?#]+)/i },
-    { fieldName: 'youtube', pattern: /(?:https?:\/\/)?(?:m\.|www\.)?(?:youtube\.com|youtu\.be)\/([^/?#]+)/i },
-    { fieldName: 'twitter', pattern: /(?:https?:\/\/)?(?:www\.)?(?:twitter\.com|x\.com)\/([^/?#]+)/i },
-    { fieldName: 'telegram', pattern: /(?:https?:\/\/)?(?:www\.)?(?:t\.me|telegram\.me|telegram\.dog)\/([^/?#]+)/i },
-  ];
-
-  for (const matcher of socialUrlMatchers) {
-    const match = trimmed.match(matcher.pattern);
-    if (!match?.[1]) continue;
-
-    const normalizedValue = String(match[1]).replace(/^@/, '').trim();
-    if (normalizedValue) {
-      return { fieldName: matcher.fieldName, value: normalizedValue };
-    }
-  }
-
-  if (/^(https?:\/\/|www\.)/i.test(trimmed) || /^[\w-]+(\.[\w-]+)+/i.test(trimmed)) {
-    return { fieldName: 'otherLink', value: normalizeUrlForStorage(trimmed) };
-  }
-
-  return null;
-};
 const REPRODUCTIVE_FIELDS = [
   'birth',
   'ownKids',
@@ -2107,23 +2062,15 @@ ${entries.join('\n')}`;
       const nextState = { ...prevState };
 
       if (!parsed) {
-        const normalizedPhone = normalizePhoneValue(rawValue);
-        const looksLikePhone =
-          Boolean(normalizedPhone) && /^[+]?\d[\d\s().-]*$/.test(rawValue);
-        const looksLikeEmail = EMAIL_REGEX.test(rawValue);
-        const resolvedTechnicalTarget = resolvePpTechnicalInputTarget(rawValue);
+        const resolvedTechnicalTarget = resolvePpTechnicalInputTarget(rawValue, {
+          normalizePhone: normalizePhoneValue,
+        });
 
         if (resolvedTechnicalTarget) {
           nextState[resolvedTechnicalTarget.fieldName] = appendFieldValue(
             prevState[resolvedTechnicalTarget.fieldName],
             resolvedTechnicalTarget.value
           );
-        } else if (looksLikeEmail) {
-          nextState.email = appendFieldValue(prevState.email, rawValue);
-        } else if (looksLikePhone) {
-          nextState.phone = appendFieldValue(prevState.phone, normalizedPhone);
-        } else {
-          nextState.name = appendFieldValue(prevState.name, rawValue);
         }
       } else {
         const { contactType, contactValues, name, surname } = parsed;
